make -dbg multi-arch: allowed now that it includes debug symbols for the library
as the library can be multi-installed through multi-arch (although not across sonames
unless I rename the debug package to libkyototycoon<so#>-dbg and have it provide:
kyototycoon-dbg. But i think that is kinda ugly until it seems like multiple versions
will want to be installed at the same time---open a bug if you want this)

as these use the new buildid naming mechanism they are totally cross-arch installable
and while I don't really any package will ever depend on them, there is no reason they
can't be same, and it seems to make more sense.